Add random sheet generation sample in MainActivity

+++ app/src/main/java/kr/blogspot/crowjdh/inspirationgen/music/models/Notable.kt
@@ -11,5 +11,5 @@ interface Notable: TickType {
 
     val length: NoteLength
 
-    override fun ticks(timeSignature: TimeSignature) = length.ticks(timeSignature.tpqn)
+    override fun ticks(timeSignature: TimeSignature) = length.ticks(timeSignature.tpqn).toLong()
 }
